+++ packages/web/src/worker/sync/SharedSyncImplementation.ts
@@ -12,6 +12,7 @@ import {
   DBAdapter,
   PowerSyncBackendConnector,
   SqliteBucketStorage,
+  SubscribedStream,
   SyncStatus
 } from '@powersync/common';
 import { Mutex } from 'async-mutex';
@@ -55,7 +56,7 @@ export type ManualSharedSyncPayload = {
  * @internal
  */
 export type SharedSyncInitOptions = {
-  streamOptions: Omit<WebStreamingSyncImplementationOptions, 'adapter' | 'uploadCrud' | 'remote'>;
+  streamOptions: Omit<WebStreamingSyncImplementationOptions, 'adapter' | 'uploadCrud' | 'remote' | 'subscriptions'>;
   dbParams: ResolvedWebSQLOpenOptions;
 };
 
@@ -73,6 +74,8 @@ export type WrappedSyncPort = {
   port: MessagePort;
   clientProvider: Comlink.Remote<AbstractSharedSyncClientProvider>;
   db?: DBAdapter;
+  currentSubscriptions: SubscribedStream[];
+  closeListeners: (() => void)[];
 };
 
 /**
@@ -94,10 +97,7 @@ const CONNECTOR_PLACEHOLDER = {} as PowerSyncBackendConnector;
  * @internal
  * Shared sync implementation which runs inside a shared webworker
  */
-export class SharedSyncImplementation
-  extends BaseObserver<SharedSyncImplementationListener>
-  implements StreamingSyncImplementation
-{
+export class SharedSyncImplementation extends BaseObserver<SharedSyncImplementationListener> {
   protected ports: WrappedSyncPort[];
 
   protected isInitialized: Promise<void>;
@@ -111,6 +111,7 @@ export class SharedSyncImplementation
   protected logger: ILogger;
   protected lastConnectOptions: PowerSyncConnectionOptions | undefined;
   protected portMutex: Mutex;
+  private subscriptions: SubscribedStream[] = [];
 
   protected connectionManager: ConnectionManager;
   syncStatus: SyncStatus;
@@ -186,6 +187,25 @@ export class SharedSyncImplementation
     return this.isInitialized;
   }
 
+  private collectActiveSubscriptions() {
+    this.logger.debug('Collecting active stream subscriptions across tabs');
+    const active = new Map<string, SubscribedStream>();
+    for (const port of this.ports) {
+      for (const stream of port.currentSubscriptions) {
+        const serializedKey = JSON.stringify(stream);
+        active.set(serializedKey, stream);
+      }
+    }
+    this.subscriptions = [...active.values()];
+    this.logger.debug('Collected stream subscriptions', this.subscriptions);
+    this.connectionManager.syncStreamImplementation?.updateSubscriptions(this.subscriptions);
+  }
+
+  updateSubscriptions(port: WrappedSyncPort, subscriptions: SubscribedStream[]) {
+    port.currentSubscriptions = subscriptions;
+    this.collectActiveSubscriptions();
+  }
+
   setLogLevel(level: ILogLevel) {
     this.logger.setLevel(level);
     this.broadCastLogger.setLevel(level);
@@ -196,6 +216,7 @@ export class SharedSyncImplementation
    */
   async setParams(params: SharedSyncInitOptions) {
     await this.portMutex.runExclusive(async () => {
+      this.collectActiveSubscriptions();
       if (this.syncParams) {
         // Cannot modify already existing sync implementation params
         // But we can ask for a DB adapter, if required, at this point.
@@ -250,11 +271,13 @@ export class SharedSyncImplementation
    * Adds a new client tab's message port to the list of connected ports
    */
   async addPort(port: MessagePort) {
-    await this.portMutex.runExclusive(() => {
+    return await this.portMutex.runExclusive(() => {
       const portProvider = {
         port,
-        clientProvider: Comlink.wrap<AbstractSharedSyncClientProvider>(port)
-      };
+        clientProvider: Comlink.wrap<AbstractSharedSyncClientProvider>(port),
+        currentSubscriptions: [],
+        closeListeners: []
+      } satisfies WrappedSyncPort;
       this.ports.push(portProvider);
 
       // Give the newly connected client the latest status
@@ -262,6 +285,8 @@ export class SharedSyncImplementation
       if (status) {
         portProvider.clientProvider.statusChanged(status.toJSON());
       }
+
+      return portProvider;
     });
   }
 
@@ -269,12 +294,12 @@ export class SharedSyncImplementation
    * Removes a message port client from this manager's managed
    * clients.
    */
-  async removePort(port: MessagePort) {
+  async removePort(port: WrappedSyncPort) {
     // Remove the port within a mutex context.
     // Warns if the port is not found. This should not happen in practice.
     // We return early if the port is not found.
     const { trackedPort, shouldReconnect } = await this.portMutex.runExclusive(async () => {
-      const index = this.ports.findIndex((p) => p.port == port);
+      const index = this.ports.findIndex((p) => p == port);
       if (index < 0) {
         this.logger.warn(`Could not remove port ${port} since it is not present in active ports.`);
         return {};
@@ -289,7 +314,7 @@ export class SharedSyncImplementation
        * not resolve. Abort them here.
        */
       [this.fetchCredentialsController, this.uploadDataController].forEach((abortController) => {
-        if (abortController?.activePort.port == port) {
+        if (abortController?.activePort == port) {
           abortController!.controller.abort(
             new AbortOperation('Closing pending requests after client port is removed')
           );
@@ -308,10 +333,13 @@ export class SharedSyncImplementation
       return () => {};
     }
 
+    for (const closeListener of trackedPort.closeListeners) {
+      closeListener();
+    }
+
     if (this.dbAdapter && this.dbAdapter == trackedPort.db) {
-      if (shouldReconnect) {
-        await this.connectionManager.disconnect();
-      }
+      // Unconditionally close the connection because the database it's writing to has just been closed.
+      await this.connectionManager.disconnect();
 
       // Clearing the adapter will result in a new one being opened in connect
       this.dbAdapter = null;
@@ -321,9 +349,9 @@ export class SharedSyncImplementation
       }
     }
 
-    if (trackedPort.db) {
-      await trackedPort.db.close();
-    }
+    // Re-index subscriptions, the subscriptions of the removed port would no longer be considered.
+    this.collectActiveSubscriptions();
+
     // Release proxy
     return () => trackedPort.clientProvider[Comlink.releaseProxy]();
   }
@@ -427,6 +455,7 @@ export class SharedSyncImplementation
         });
       },
       ...syncParams.streamOptions,
+      subscriptions: this.subscriptions,
       // Logger cannot be transferred just yet
       logger: this.logger
     });
@@ -445,11 +474,21 @@ export class SharedSyncImplementation
     const locked = new LockedAsyncDatabaseAdapter({
       name: identifier,
       openConnection: async () => {
-        return new WorkerWrappedAsyncDatabaseConnection({
+        const wrapped = new WorkerWrappedAsyncDatabaseConnection({
           remote,
           baseConnection: db,
-          identifier
+          identifier,
+          // It's possible for this worker to outlive the client hosting the database for us. We need to be prepared for
+          // that and ensure pending requests are aborted when the tab is closed.
+          remoteCanCloseUnexpectedly: true
         });
+        lastClient.closeListeners.push(() => {
+          this.logger.info('Aborting open connection because associated tab closed.');
+          wrapped.close();
+          wrapped.markRemoteClosed();
+        });
+
+        return wrapped;
       },
       logger: this.logger
     });
@@ -470,7 +509,7 @@ export class SharedSyncImplementation
    * A function only used for unit tests which updates the internal
    * sync stream client and all tab client's sync status
    */
-  private async _testUpdateAllStatuses(status: SyncStatusOptions) {
+  async _testUpdateAllStatuses(status: SyncStatusOptions) {
     if (!this.connectionManager.syncStreamImplementation) {
       throw new Error('Cannot update status without a sync stream implementation');
     }