+++ packages/react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @powersync/react
 
+## 1.8.0
+
+### Minor Changes
+
+- 7a5aaf5: Add hooks for sync streams
+
+### Patch Changes
+
+- d8236aa: Fix "order and size of this array must remain constant" warning.
+- Updated dependencies [eff8cbf]
+  - @powersync/common@1.39.0
+
+## 1.7.4
+
+### Patch Changes
+
+- a0ee132: - Fixed bug where the `useQuery` reported `error` state would not clear after updating the query to a valid query.
+  - Fixed bug where `useQuery` `isFetching` status would not immediately be reported as true when the query has changed.
+- Updated dependencies [a0ee132]
+- Updated dependencies [ba72a58]
+  - @powersync/common@1.38.1
+
 ## 1.7.3
 
 ### Patch Changes