+++ packages/web/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @powersync/web
 
+## 1.27.0
+
+### Minor Changes
+
+- eff8cbf: Add alpha support for sync streams, allowing different sets of data to be synced dynamically.
+
+### Patch Changes
+
+- c9c1e24: Fix issues with multiple tabs when the Rust client and OPFS is used.
+- c78071f: Update core extension to 0.4.6
+- Updated dependencies [eff8cbf]
+  - @powersync/common@1.39.0
+
+## 1.26.2
+
+### Patch Changes
+
+- Updated dependencies [a0ee132]
+- Updated dependencies [ba72a58]
+  - @powersync/common@1.38.1
+
 ## 1.26.1
 
 ### Patch Changes